refactor: move json schema parser helper function to utils file
It's better suited to be there, for now it's only dont int the
server project
